# docs/apidocs/thorapi_openapi_v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Thor.Api | v1",
    "version": "1.0"
  },
  "paths": {
...
    "/v1/authentication-methods/types": {
      "get": {
        "tags": [
          "AuthenticationMethods"
        ],
        "parameters": [
          {
            "name": "connectorType",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int16"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthenticationTypeResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthenticationTypeResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthenticationTypeResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/authentication-methods/types/{typeId}/fields": {
      "get": {
        "tags": [
          "AuthenticationMethods"
        ],
        "parameters": [
          {
            "name": "typeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthenticationFieldResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthenticationFieldResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthenticationFieldResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/identity-providers/{providerName}": {
      "get": {
        "tags": [
          "IdentityProviders"
        ],
        "parameters": [
          {
            "name": "providerName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TenantIdentityProviderResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantIdentityProviderResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantIdentityProviderResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "IdentityProviders"
        ],
        "parameters": [
          {
            "name": "providerName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantIdentityProviderRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantIdentityProviderRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantIdentityProviderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TenantIdentityProviderResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantIdentityProviderResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantIdentityProviderResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "IdentityProviders"
        ],
        "parameters": [
          {
            "name": "providerName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthenticationFieldResponse": {
        "required": [
          "id",
          "name",
          "displayName",
          "inputType",
          "description"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "inputType": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "AuthenticationTypeResponse": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "OidcProviderSettings": {
        "required": [
          "clientId",
          "clientSecret",
          "issuer",
          "authorizeScopes",
          "attributesRequestMethod"
        ],
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string"
          },
          "clientSecret": {
            "type": "string"
          },
          "issuer": {
            "type": "string"
          },
          "authorizeScopes": {
            "type": "string"
          },
          "attributesRequestMethod": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "OidcProviderSettingsResponse": {
        "required": [
          "clientId",
          "issuer",
          "authorizeScopes",
          "attributesRequestMethod"
        ],
        "type": "object",
        "properties": {
          "clientId": {
            "type": [
              "null",
              "string"
            ]
          },
          "issuer": {
            "type": [
              "null",
              "string"
            ]
          },
          "authorizeScopes": {
            "type": [
              "null",
              "string"
            ]
          },
          "attributesRequestMethod": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SamlProviderSettings": {
        "required": [
          "metadataUrl",
          "metadataFile",
          "idpSignout"
        ],
        "type": "object",
        "properties": {
          "metadataUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "metadataFile": {
            "type": [
              "null",
              "string"
            ]
          },
          "idpSignout": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
      "SamlProviderSettingsResponse": {
        "required": [
          "metadataUrl",
          "idpSignout"
        ],
        "type": "object",
        "properties": {
          "metadataUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "idpSignout": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
...
      "TenantIdentityProviderResponse": {
        "required": [
          "providerName",
          "providerType",
          "createdAt",
          "updatedAt",
          "saml",
          "oidc",
          "attributeMapping",
          "idpIdentifiers"
        ],
        "type": "object",
        "properties": {
          "providerName": {
            "type": "string"
          },
          "providerType": {
            "type": "string"
          },
          "createdAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "updatedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "saml": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SamlProviderSettingsResponse"
              }
            ]
          },
          "oidc": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OidcProviderSettingsResponse"
              }
            ]
          },
          "attributeMapping": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "idpIdentifiers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "UpdateTenantIdentityProviderRequest": {
        "required": [
          "providerType",
          "saml",
          "oidc",
          "attributeMapping",
          "idpIdentifiers"
        ],
        "type": "object",
        "properties": {
          "providerType": {
            "type": "string"
          },
          "saml": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SamlProviderSettings"
              }
            ]
          },
          "oidc": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OidcProviderSettings"
              }
            ]
          },
          "attributeMapping": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "string"
            }
          },
          "idpIdentifiers": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  },
...
}